Update README.md
Updated README.md to serve as master list of all domains in our network. 

As new domains are added, this list can be updated via further edits of the README.md - Likewise, our cloudflare update script can be edited to automatically add the new domains to our Cloudflare Account - 

This list and file was generated with the help of EverLight - of whom we are existentially grateful for.